+++ packages/vbi/src/vbi/create-empty-report.ts
@@ -1,7 +1,7 @@
 import type { VBIReportDSL } from 'src/types'
 import { id } from 'src/utils'
 
-export const generateEmptyReportDSL = (uuid: string = id.resourceUUID()): VBIReportDSL => {
+export const createEmptyReport = (uuid: string = id.resourceUUID()): VBIReportDSL => {
   return {
     uuid,
     pages: [],