+++ packages/vbi/src/vbi/create-vbi.ts
@@ -7,18 +7,18 @@ import type {
   VBIReportBuilderOptions,
   VBIReportDSLInput,
 } from 'src/types'
-import { createChartBuilderFromVBIChartDSLInput } from './from/from-vbi-dsl-input'
-import { createInsightBuilderFromVBIInsightDSLInput } from './from/from-vbi-insight-dsl-input'
-import { createReportBuilderFromVBIReportDSLInput } from './from/from-vbi-report-dsl-input'
-import { generateEmptyChartDSL } from './generate-empty-dsl'
-import { generateEmptyInsightDSL } from './generate-empty-insight-dsl'
-import { generateEmptyReportDSL } from './generate-empty-report-dsl'
-import { generateEmptyReportPageDSL } from './generate-empty-report-page-dsl'
+import {
+  createChartBuilderFromVBIChartDSLInput,
+  createInsightBuilderFromVBIInsightDSLInput,
+  createReportBuilderFromVBIReportDSLInput,
+} from './from'
+import { createVBIResourceRegistry } from './resources'
+import { createEmptyChart } from './create-empty-chart'
+import { createEmptyInsight } from './create-empty-insight'
+import { createEmptyReport } from './create-empty-report'
+import { createEmptyReportPage } from './create-empty-report-page'
 import { mergeBuilderOptions, mergeReportBuilderOptions } from './merge-builder-options'
-import { createVBIResourceRegistry } from './resource-registry'
 import type { VBIInstance } from './types'
-
-export type { VBIInstance } from './types'
 
 export function createVBI(): VBIInstance<DefaultVBIQueryDSL, DefaultVBISeedDSL>
 export function createVBI<TQueryDSL, TSeedDSL>(
@@ -27,17 +27,17 @@ export function createVBI<TQueryDSL, TSeedDSL>(
 export function createVBI<TQueryDSL = DefaultVBIQueryDSL, TSeedDSL = DefaultVBISeedDSL>(
   defaultBuilderOptions?: VBIChartBuilderOptions<TQueryDSL, TSeedDSL>,
 ) {
-  const resourceRegistry = createVBIResourceRegistry()
+  const resourceRegistry = createVBIResourceRegistry<TQueryDSL, TSeedDSL>()
 
   const createChart = (vbi: VBIChartDSLInput, builderOptions?: VBIChartBuilderOptions<TQueryDSL, TSeedDSL>) => {
     const options = mergeBuilderOptions(defaultBuilderOptions, builderOptions)
     const builder = createChartBuilderFromVBIChartDSLInput(vbi, options)
-    resourceRegistry.charts.set(builder.getUUID(), builder)
+    resourceRegistry.charts.registerBuilder(builder.getUUID(), builder)
     return builder
   }
   const createInsight = (insight: VBIInsightDSLInput) => {
     const builder = createInsightBuilderFromVBIInsightDSLInput(insight)
-    resourceRegistry.insights.set(builder.getUUID(), builder)
+    resourceRegistry.insights.registerBuilder(builder.getUUID(), builder)
     return builder
   }
   const createReport = (report: VBIReportDSLInput, builderOptions?: VBIReportBuilderOptions<TQueryDSL, TSeedDSL>) => {
@@ -51,16 +51,16 @@ export function createVBI<TQueryDSL = DefaultVBIQueryDSL, TSeedDSL = DefaultVBIS
     getConnector,
     chart: {
       create: createChart,
-      generateEmptyDSL: generateEmptyChartDSL,
+      createEmpty: createEmptyChart,
     },
     insight: {
       create: createInsight,
-      generateEmptyDSL: generateEmptyInsightDSL,
+      createEmpty: createEmptyInsight,
     },
     report: {
       create: createReport,
-      generateEmptyDSL: generateEmptyReportDSL,
-      generateEmptyPageDSL: generateEmptyReportPageDSL,
+      createEmpty: createEmptyReport,
+      createEmptyPage: createEmptyReportPage,
     },
   } satisfies VBIInstance<TQueryDSL, TSeedDSL>
 }