+++ packages/vbi/src/vbi/from/from-vbi-report-dsl-input.ts
@@ -3,13 +3,13 @@ import type { DefaultVBIQueryDSL, DefaultVBISeedDSL } from 'src/chart-builder/ad
 import type { VBIReportDSLInput, VBIReportBuilderOptions } from 'src/types'
 import { zVBIReportDSL } from 'src/types/reportDSL/report'
 import { VBIReportBuilder } from 'src/report-builder/builder'
-import type { VBIResourceRegistry } from '../resource-registry'
+import type { VBIResourceRegistry } from '../resources'
 import { ensureReportPages } from './report-page-y-map'
 
 export const createReportBuilderFromVBIReportDSLInput = <TQueryDSL = DefaultVBIQueryDSL, TSeedDSL = DefaultVBISeedDSL>(
   report: VBIReportDSLInput,
   options?: VBIReportBuilderOptions<TQueryDSL, TSeedDSL>,
-  resourceRegistry?: VBIResourceRegistry,
+  resourceRegistry?: VBIResourceRegistry<TQueryDSL, TSeedDSL>,
 ) => {
   const doc = new Y.Doc()
   const dsl = doc.getMap('dsl')