+++ practices/minimalist/src/utils/demoConnector.ts
@@ -63,5 +63,5 @@ export const registerDemoConnector = () => {
 
 registerDemoConnector();
 export const defaultBuilder: VBIChartBuilder = VBI.chart.create(
-  VBI.chart.generateEmptyDSL(connectorId),
+  VBI.chart.createEmpty(connectorId),
 );